# Optimizations of kNN BF prediction for large k 