--- a/cpp/daal/src/algorithms/k_nearest_neighbors/bf_knn_impl.i
+++ b/cpp/daal/src/algorithms/k_nearest_neighbors/bf_knn_impl.i
@@ -35,6 +35,7 @@
 #include "src/algorithms/service_kernel_math.h"
 #include "src/algorithms/service_sort.h"
 #include "src/externals/service_math.h"
+#include "src/algorithms/k_nearest_neighbors/knn_heap.h"
 
 namespace daal
 {
@@ -51,6 +52,9 @@ public:
     BruteForceNearestNeighbors() {}
 
     ~BruteForceNearestNeighbors() {}
+
+    typedef GlobalNeighbors<FPType, cpu> Neighbors;
+    typedef MinHeap<Neighbors, cpu> HeapType;
 
     services::Status kNeighbors(const size_t k, const size_t nClasses, VoteWeights voteWeights, DAAL_UINT64 resultsToCompute,
                                 DAAL_UINT64 resultsToEvaluate, const NumericTable * trainTable, const NumericTable * testTable,
@@ -85,6 +89,8 @@ public:
         TlsMem<int, cpu> tlsKIndexes(inBlockSize * k);
         TlsMem<FPType, cpu> tlsVoting(nClasses);
 
+        TlsMem<HeapType, cpu> tlsHeap(outBlockSize);
+
         SafeStatus safeStat;
 
         daal::threader_for(nOuterBlocks, nOuterBlocks, [&](size_t outerBlock) {
@@ -94,12 +100,68 @@ public:
 
             DAAL_CHECK_STATUS_THR(computeKNearestBlock(&euclDist, outerSize, inBlockSize, outerStart, nTrain, resultsToEvaluate, resultsToCompute,
                                                        nClasses, k, voteWeights, trainLabel, trainTable, testTable, testLabelTable, indicesTable,
-                                                       distancesTable, tlsDistances, tlsIdx, tlsKDistances, tlsKIndexes, tlsVoting, nOuterBlocks));
+                                                       distancesTable, tlsDistances, tlsIdx, tlsKDistances, tlsKIndexes, tlsVoting, tlsHeap, nOuterBlocks));
         });
 
         if (resultsToEvaluate & daal::algorithms::classifier::computeClassLabels)
         {
             newTrainLabelTable->releaseBlockOfRows(trainLabelBlock);
+        }
+
+        printf("\nLabels:\n");
+        if (resultsToEvaluate & daal::algorithms::classifier::computeClassLabels)
+        {
+            daal::internal::ReadRows<int, cpu> testLabelRows(testLabelTable, 0, nTest);
+
+            printf("%d ", testLabelRows.get()[0]);
+            printf("\n");
+            printf("%d ", testLabelRows.get()[nTest / 2]);
+            printf("\n");
+            printf("%d ", testLabelRows.get()[(nTest - 1) ]);
+            printf("\n");
+        }
+
+        printf("Distance:\n");
+        if (resultsToCompute & computeDistances)
+        {
+            daal::internal::ReadRows<FPType, cpu> distancesBlock(distancesTable, 0, nTest);
+
+            for (size_t i = 0; i < k; ++i)
+            {
+                printf("%f ", distancesBlock.get()[0 * k + i]);
+            }
+            printf("\n");
+            for (size_t i = 0; i < k; ++i)
+            {
+                printf("%f ", distancesBlock.get()[nTest / 2 * k + i]);
+            }
+            printf("\n");
+            for (size_t i = 0; i < k; ++i)
+            {
+                printf("%f ", distancesBlock.get()[(nTest - 1) * k + i]);
+            }
+            printf("\n");
+        }
+        printf("Indexes:\n");
+        if (resultsToCompute & computeIndicesOfNeighbors)
+        {
+            daal::internal::ReadRows<int, cpu> indexesBlock(indicesTable, 0, nTest);
+
+            for (size_t i = 0; i < k; ++i)
+            {
+                printf("%d ", indexesBlock.get()[0 * k + i]);
+            }
+            printf("\n");
+            for (size_t i = 0; i < k; ++i)
+            {
+                printf("%d ", indexesBlock.get()[nTest / 2 * k + i]);
+            }
+            printf("\n");
+            for (size_t i = 0; i < k; ++i)
+            {
+                printf("%d ", indexesBlock.get()[(nTest - 1) * k + i]);
+            }
+            printf("\n");
         }
 
         return safeStat.detach();
@@ -110,9 +172,8 @@ protected:
     {
     public:
         DAAL_NEW_DELETE();
-        FPType * kDistances;
         FPType * maxs;
-        int * kIndexes;
+        HeapType* heapsData;
 
         static BruteForceTask * create(const size_t inBlockSize, const size_t outBlockSize, const size_t k)
         {
@@ -122,26 +183,26 @@ protected:
             return nullptr;
         }
 
-        bool isValid() const { return _buff.get(); }
+        bool isValid() const { return _buff.get() && _heaps.get(); }
 
     private:
         BruteForceTask(size_t inBlockSize, size_t outBlockSize, size_t k)
         {
-            const size_t kDistancesSize = inBlockSize * k * sizeof(FPType);
-            const size_t maxsSize       = inBlockSize * sizeof(FPType);
-            const size_t kIndexesSize   = inBlockSize * k * sizeof(int);
-
-            _buff.reset(kDistancesSize + maxsSize + kIndexesSize);
-
-            kDistances = reinterpret_cast<FPType *>(&_buff[0]);
-            maxs       = reinterpret_cast<FPType *>(&_buff[kDistancesSize]);
-            kIndexes   = reinterpret_cast<int *>(&_buff[kDistancesSize + maxsSize]);
-
+            _buff.reset(outBlockSize);
+            maxs       = _buff.get();
             service_memset_seq<FPType, cpu>(maxs, MaxVal<FPType>::get(), outBlockSize);
-            service_memset_seq<FPType, cpu>(kDistances, MaxVal<FPType>::get(), inBlockSize * k);
+
+            _heaps.reset(outBlockSize);
+
+            for (size_t i = 0; i < outBlockSize; ++i)
+            {
+                _heaps[i].init(k);
+            }
+            heapsData = _heaps.get();
         }
 
-        TArrayScalable<byte, cpu> _buff;
+        TArrayScalable<FPType, cpu> _buff;
+        TArrayScalable<HeapType, cpu> _heaps;
     };
 
     services::Status computeKNearestBlock(daal::algorithms::internal::EuclideanDistances<FPType, cpu> * distancesInstance, const size_t blockSize,
@@ -150,7 +211,8 @@ protected:
                                           FPType * trainLabel, const NumericTable * trainTable, const NumericTable * testTable,
                                           NumericTable * testLabelTable, NumericTable * indicesTable, NumericTable * distancesTable,
                                           TlsMem<FPType, cpu> & tlsDistances, TlsMem<int, cpu> & tlsIdx, TlsMem<FPType, cpu> & tlsKDistances,
-                                          TlsMem<int, cpu> & tlsKIndexes, TlsMem<FPType, cpu> & tlsVoting, size_t nOuterBlocks)
+                                          TlsMem<int, cpu> & tlsKIndexes, TlsMem<FPType, cpu> & tlsVoting,
+                                          TlsMem<HeapType, cpu> & tlsHeap, size_t nOuterBlocks)
     {
         const size_t inBlockSize = trainBlockSize;
         const size_t inRows      = nTrain;
@@ -167,7 +229,7 @@ protected:
         SafeStatus safeStat;
 
         daal::tls<BruteForceTask *> tlsTask([=, &safeStat]() {
-            auto tlsData = BruteForceTask::create(inBlockSize, blockSize, k);
+            auto tlsData = BruteForceTask::create(inBlockSize, iSize, k);
             if (!tlsData)
             {
                 safeStat.add(services::ErrorMemoryAllocationFailed);
@@ -175,7 +237,9 @@ protected:
             return tlsData;
         });
 
-        daal::conditional_threader_for(nOuterBlocks > 3 * threader_get_max_threads_number(), nInBlocks, [&](size_t inBlock) {
+        const size_t nThreads = _daal_threader_get_max_threads();
+        daal::conditional_threader_for(nOuterBlocks < 2 * nThreads, nInBlocks, [&](size_t inBlock)
+        {
             const size_t j1    = inBlock * inBlockSize;
             const size_t j2    = (inBlock + 1 == nInBlocks ? inRows : j1 + inBlockSize);
             const size_t jSize = j2 - j1;
@@ -190,8 +254,7 @@ protected:
             DAAL_CHECK_MALLOC_THR(idx);
 
             FPType * maxs            = tls->maxs;
-            FPType * kDistancesLocal = tls->kDistances;
-            int * kIndexesLocal      = tls->kIndexes;
+            HeapType* heapsLocal      = tls->heapsData;
 
             ReadRows<FPType, cpu> outDataRows(const_cast<NumericTable *>(trainTable), j1, j2 - j1);
             DAAL_CHECK_BLOCK_STATUS_THR(outDataRows);
@@ -205,7 +268,7 @@ protected:
 
                 if (indexes)
                 {
-                    updateLocalNeighbours(indexes, idx, jSize, i, k, kDistancesLocal, kIndexesLocal, maxs, distancesBuff, j1);
+                    updateLocalNeighbours(indexes, idx, jSize, i, k, maxs, distancesBuff, j1, heapsLocal[i]);
                 }
             }
         });
@@ -216,52 +279,73 @@ protected:
         FPType * kDistances = tlsKDistances.local();
         DAAL_CHECK_MALLOC(kDistances);
 
-        service_memset_seq<FPType, cpu>(kDistances, MaxVal<FPType>::get(), blockSize * k);
+        // TODO
+        // HeapType* heaps = tlsHeap.local();
+        // DAAL_CHECK_MALLOC(heaps);
+
+        TArrayScalable<HeapType, cpu> heaps(iSize);
+
+        for (size_t i = 0; i < iSize; ++i)
+        {
+            heaps[i].init(k);
+        }
 
         tlsTask.reduce([&](BruteForceTask * tls) {
-            FPType * kDistancesLocal = tls->kDistances;
-            int * kIndexesLocal      = tls->kIndexes;
-
+            HeapType* heapsLocal = tls->heapsData;
             for (size_t i = 0; i < iSize; i++)
             {
-                mergeNeighbours(i, k, kIndexesLocal, kDistancesLocal, kIndexes, kDistances);
+                const size_t size = heapsLocal[i].size();
+                for (size_t j = 0; j < size; ++j)
+                {
+                    heaps[i].replaceMaxIfNeeded(heapsLocal[i][j], k);
+                }
             }
 
             delete tls;
         });
 
+        for (size_t i = 0; i < iSize; i++)
+        {
+            for (size_t kk = 0; kk < k; ++kk)
+            {
+                // max(0, d) to remove negative distances before Sqrt
+                kDistances[i * k + kk] = services::internal::max<cpu, FPType>(FPType(0), heaps[i][kk].distance);
+                kIndexes[i * k + kk]   = heaps[i][kk].index;
+            }
+        }
+
         // Euclidean Distances are computed without Sqrt, fixing it here
-        Math<FPType, cpu>::vSqrt(blockSize * k, kDistances, kDistances);
+        Math<FPType, cpu>::vSqrt(iSize * k, kDistances, kDistances);
 
         // sort by distances
-        for (size_t i = 0; i < blockSize; ++i)
+        for (size_t i = 0; i < iSize; ++i)
         {
             daal::algorithms::internal::qSort<FPType, int, cpu>(k, kDistances + i * k, kIndexes + i * k);
         }
 
         if (resultsToCompute & computeIndicesOfNeighbors)
         {
-            daal::internal::WriteOnlyRows<int, cpu> indexesBlock(indicesTable, startTestIdx, blockSize);
+            daal::internal::WriteOnlyRows<int, cpu> indexesBlock(indicesTable, startTestIdx, iSize);
             DAAL_CHECK_BLOCK_STATUS(indexesBlock);
             int * indices = indexesBlock.get();
 
-            const size_t size = blockSize * k * sizeof(*indices);
+            const size_t size = iSize * k * sizeof(*indices);
             daal::services::internal::daal_memcpy_s(indices, size, kIndexes, size);
         }
 
         if (resultsToCompute & computeDistances)
         {
-            daal::internal::WriteOnlyRows<FPType, cpu> distancesBlock(distancesTable, startTestIdx, blockSize);
+            daal::internal::WriteOnlyRows<FPType, cpu> distancesBlock(distancesTable, startTestIdx, iSize);
             DAAL_CHECK_BLOCK_STATUS(distancesBlock);
             FPType * distances = distancesBlock.get();
 
-            const size_t size = blockSize * k * sizeof(FPType);
+            const size_t size = iSize * k * sizeof(FPType);
             daal::services::internal::daal_memcpy_s(distances, size, kDistances, size);
         }
 
         if (resultsToEvaluate & daal::algorithms::classifier::computeClassLabels)
         {
-            daal::internal::WriteOnlyRows<int, cpu> testLabelRows(testLabelTable, startTestIdx, blockSize);
+            daal::internal::WriteOnlyRows<int, cpu> testLabelRows(testLabelTable, startTestIdx, iSize);
             DAAL_CHECK_BLOCK_STATUS(testLabelRows);
             int * testLabel = testLabelRows.get();
 
@@ -270,11 +354,11 @@ protected:
 
             if (voteWeights == VoteWeights::voteUniform)
             {
-                DAAL_CHECK_STATUS_VAR(uniformWeightedVoting(nClasses, k, blockSize, nTrain, kIndexes, trainLabel, testLabel, voting));
+                DAAL_CHECK_STATUS_VAR(uniformWeightedVoting(nClasses, k, iSize, nTrain, kIndexes, trainLabel, testLabel, voting));
             }
             else
             {
-                DAAL_CHECK_STATUS_VAR(distanceWeightedVoting(nClasses, k, blockSize, nTrain, kDistances, kIndexes, trainLabel, testLabel, voting));
+                DAAL_CHECK_STATUS_VAR(distanceWeightedVoting(nClasses, k, iSize, nTrain, kDistances, kIndexes, trainLabel, testLabel, voting));
             }
         }
 
@@ -295,59 +379,21 @@ protected:
         return count;
     }
 
-    void updateLocalNeighbours(size_t indexes, int * idx, size_t jSize, size_t i, size_t k, FPType * kDistances, int * kIndexes, FPType * maxs,
-                               FPType * distances, size_t j1)
+    void updateLocalNeighbours(size_t indexes, int * idx, size_t jSize, size_t i, size_t k, FPType * maxs,
+                               FPType * distances, size_t j1, HeapType& heap)
     {
         for (size_t j = 0; j < indexes; ++j)
         {
             FPType d = distances[i * jSize + idx[j]];
 
-            int maxIdx = 0;
-            for (size_t kk = 0; kk < k; ++kk)
-            {
-                if (d < kDistances[i * k + kk] && kDistances[i * k + kk] > kDistances[i * k + maxIdx])
-                {
-                    maxIdx = kk;
-                }
-            }
-            if (kDistances[i * k + maxIdx] > d)
-            {
-                kDistances[i * k + maxIdx] = d;
-                kIndexes[i * k + maxIdx]   = idx[j] + j1;
-            }
+            Neighbors neigh;
+            neigh.distance = d;
+            neigh.index    = idx[j] + j1;
+
+            heap.replaceMaxIfNeeded(neigh, k);
         }
 
-        FPType max = kDistances[i * k + 0];
-        for (size_t kk = 1; kk < k; ++kk)
-        {
-            if (kDistances[i * k + kk] > max)
-            {
-                max = kDistances[i * k + kk];
-            }
-        }
-        maxs[i] = max;
-    }
-
-    void mergeNeighbours(size_t i, size_t k, int * idxLocal, FPType * distancesLocal, int * kIndexes, FPType * kDistances)
-    {
-        for (size_t j = 0; j < k; ++j)
-        {
-            FPType d = distancesLocal[i * k + j];
-
-            int maxIdx = 0;
-            for (size_t kk = 0; kk < k; ++kk)
-            {
-                if (d < kDistances[i * k + kk] && kDistances[i * k + kk] > kDistances[i * k + maxIdx])
-                {
-                    maxIdx = kk;
-                }
-            }
-            if (kDistances[i * k + maxIdx] > d)
-            {
-                kDistances[i * k + maxIdx] = d;
-                kIndexes[i * k + maxIdx]   = idxLocal[i * k + j];
-            }
-        }
+        maxs[i] = heap.getMax()->distance;
     }
 
     services::Status uniformWeightedVoting(const size_t nClasses, const size_t k, const size_t n, const size_t nTrain, int * indices,