FP | n | p | k | Before, ms | After, ms | Speedup
-- | -- | -- | -- | -- | -- | --
double | 10000 | 10 | 5 | 6.4 | 5.9 | 1.1
double | 10000 | 10 | 1000 | 5051 | 191 | 26.5
single | 10000 | 10 | 5 | 4.7 | 3.7 | 1.3
single | 10000 | 10 | 1000 | 4868 | 178 | 27.4
double | 10000 | 1000 | 5 | 269 | 268 | 1.0
double | 10000 | 1000 | 1000 | 5345 | 530 | 10.1
single | 10000 | 1000 | 5 | 141 | 143 | 1.0
single | 10000 | 1000 | 1000 | 5032 | 364 | 13.8

Also:
- fixing an issues with incorrect results due to NaNs in distances
- fixing an issue with conditional threading for small test data

CC @owerbat, @Alexsandruss