@owerbat 2020-10-19
I've checked, everything ok with scikit-learn patching. Tests are passed and I see nice speedup with large number of neighbors